+++ src/Cooked/MockChain.hs
@@ -2,17 +2,16 @@
 -- elements related to logs and inner state.
 module Cooked.MockChain (module X) where
 
-import Cooked.MockChain.AutoFilling as X
-import Cooked.MockChain.Balancing as X
+import Cooked.MockChain.Automation.Balancing as X
 import Cooked.MockChain.Common as X
-import Cooked.MockChain.Error as X
-import Cooked.MockChain.Instances as X
-import Cooked.MockChain.Journal as X
-import Cooked.MockChain.Misc as X
-import Cooked.MockChain.Read as X
-import Cooked.MockChain.Runnable as X
-import Cooked.MockChain.State as X
+import Cooked.MockChain.Effect.Misc as X
+import Cooked.MockChain.Effect.Read as X
+import Cooked.MockChain.Effect.Write as X
+import Cooked.MockChain.Run.Instances as X
+import Cooked.MockChain.Run.Runnable as X
+import Cooked.MockChain.Run.Tweak as X
+import Cooked.MockChain.Runtime.Error as X
+import Cooked.MockChain.Runtime.Journal as X
+import Cooked.MockChain.Runtime.State as X
 import Cooked.MockChain.Testing as X
-import Cooked.MockChain.Tweak as X
 import Cooked.MockChain.UtxoSearch as X
-import Cooked.MockChain.Write as X