+++ src/Cooked/MockChain/Automation/GenerateTx/Collateral.hs
@@ -1,11 +1,14 @@
 -- | This module exposes the generation of transaction collaterals, which
 -- consist of a collateral amount, collateral inputs and return collateral
-module Cooked.MockChain.GenerateTx.Collateral where
+module Cooked.MockChain.Automation.GenerateTx.Collateral
+  ( toCollateralTriplet,
+  )
+where
 
 import Cardano.Api qualified as Cardano
+import Cooked.MockChain.Automation.GenerateTx.Output
 import Cooked.MockChain.Common
-import Cooked.MockChain.GenerateTx.Output
-import Cooked.MockChain.Read
+import Cooked.MockChain.Effect.Read
 import Cooked.Skeleton.Output
 import Cooked.Skeleton.Value
 import Data.Map qualified as Map