+++ src/Cooked/MockChain/Testing.hs
@@ -2,16 +2,97 @@
 
 -- | This modules provides primitives to run tests over mockchain executions and
 -- to provide requirements on the the number and results of these runs.
-module Cooked.MockChain.Testing where
+module Cooked.MockChain.Testing
+  ( -- * Common interface between HUnit and QuickCheck
+    IsProp (..),
+    testBool,
+    testBoolMsg,
+    testAll,
+    testAny,
+    (.==.),
+    (.&&.),
+    (.||.),
+    assertionToMaybe,
+
+    -- * Extra HUnit assertions
+    forAll,
+    assertSubset,
+    assertSameSets,
+
+    -- * Data structure to test mockchain traces
+    FailureProp,
+    SuccessProp,
+    SizeProp,
+    LogProp,
+    StateProp,
+    Runner,
+    Test (..),
+    testToProp,
+
+    -- * Simple test templates
+    testCooked,
+    testCookedFromInitDistTemplate,
+    testCookedQC,
+    testCookedQCFromInitDistTemplate,
+    mustSucceedTest',
+    mustSucceedTest,
+    mustFailTest',
+    mustFailTest,
+
+    -- * Appending elements (in particular requirements) to existing tests
+    withInitDist,
+    withPrettyOpts,
+    withLogProp,
+    withStateProp,
+    withSuccessProp,
+    withResultProp,
+    withSizeProp,
+    withFailureProp,
+    withErrorProp,
+
+    -- * Specific properties around failures
+    isPhase1Failure,
+    isPhase2Failure,
+    isPhase1FailureWithMsg,
+    isPhase2FailureWithMsg,
+
+    -- * Specific properties around number of outcomes
+    isOfSize,
+    isAtLeastOfSize,
+    isAtMostOfSize,
+
+    -- * Specific properties over the log
+    happened,
+    didNotHappen,
+
+    -- * Specific properties over successes
+    isAtAddress,
+    possesses,
+
+    -- * Advanced test templates
+    mustFailInPhase2Test',
+    mustFailInPhase2Test,
+    mustFailInPhase2WithMsgTest',
+    mustFailInPhase2WithMsgTest,
+    mustFailInPhase1Test',
+    mustFailInPhase1Test,
+    mustFailInPhase1WithMsgTest',
+    mustFailInPhase1WithMsgTest,
+    mustSucceedWithSizeTest',
+    mustSucceedWithSizeTest,
+    mustFailWithSizeTest',
+    mustFailWithSizeTest,
+  )
+where
 
 import Control.Exception qualified as E
 import Control.Monad
-import Cooked.MockChain.Error
-import Cooked.MockChain.Journal
-import Cooked.MockChain.Log
-import Cooked.MockChain.Runnable
-import Cooked.MockChain.State
-import Cooked.MockChain.Write
+import Cooked.MockChain.Effect.Log
+import Cooked.MockChain.Effect.Write
+import Cooked.MockChain.Run.Runnable
+import Cooked.MockChain.Runtime.Error
+import Cooked.MockChain.Runtime.Journal
+import Cooked.MockChain.Runtime.State
 import Cooked.Pretty
 import Data.Default
 import Data.List (isInfixOf)